Troubleshoot guide for requested resource not found
This adds a section in troubleshoot guide on what "requested resource
not found" error means and how to proceed with it. Generally this is an
indicator that the kubectl crossplane plugin needs to be upgraded to
newer version.

Signed-off-by: Khosrow Moossavi <khos2ow@gmail.com>
(cherry picked from commit bb8cf4db57d242b82df1791c922e12c1670ed9c5)