Add FromMultipleCompositeFieldPaths patch type
We add a patch type that takes a list of composite field paths.
When applied, each field path value is retrieved from the composite
resource.

Transforms are modified to be able to accept a slice of interfaces, to
support multiple input values.

When a Transform is applied to a patch with multiple input fields, each
input field will be transformed separately.

Currently, only the first input value will be patched into the output
field. Further functionality to support combining input fields will
follow.

Signed-off-by: Ben Agricola <bagricola@squiz.co.uk>