Add support for required fromFieldPath patches
Current a fromFieldPath patch is a no-op if the specified 'from field path' does
not actually exist. This commit allows such patches to be marked as 'Required',
causing them to return an error if the field path they're patching from does not
(yet) exist. This can be used to block the creation of a composed resource until
the field path exists on the composite resource.

Signed-off-by: Nic Cope <negz@rk0n.org>