Apply composed resources before processing them
This ensures we always apply all of our composed resources before we 'process'
them - i.e. before we attempt to derive XR fields, connection details, or
readiness by observing composed resources.

Another subtle change here is that when processing we now:

1. Derive XR fields from the composed resource.
2. Derive XR connection details from the composed resource.
3. Derive XR readiness from the composed resource.

Previously the order was details, readiness, then fields. This change attempts
to ensure we'll never consider an XR to be ready if we're failing to read
connection details or to patch the XR.

Signed-off-by: Nic Cope <negz@rk0n.org>