Switch from using nonroot as distroless static base image
Changes the crossplane base image to distroless static to workaround
issues with updated runc which executes chdir to cwd later in the init
sequence after setting up the container user.

The distroless nonroot image sets user as nonroot (uid=65532) and cwd to
/home/nonroot with 0700 permissions. If the container user is not 65532
the chdir to /home/nonroot fails with permission denied, effectively
making it impossible to run as any other user. We can mostly achieve our
desired outcome of running as nonroot user, but introducing more
flexibility for platforms, such as Openshift, which assign uid ranges
for namespaces, by switching to the regular distroless static base image
and assigning runAsUser: 65532 in our helm deployment. This will
eliminate the chdir permission denied errors when a different user is
selected because the default cwd is /, which has 0755 permissions.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>