Set user to 65532 in Dockerfile
Changes default uid to 65532, which is nonroot in distroless static
base image. In practice, this gets overridden by default in helm chart
deployment, but it is helpful for them to match up if someone is only
looking at Dockerfile, and it would also cause running as nonroot user
if someone was to run outside context of deployment.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>