Add multi-patch-source policy
Add usage of new IsOptionalFieldPathNotFound to multiple-source patch.
This allows the same Policy as a standard fromFieldPath patch to be used
to control further processing of a resource. If any source fields in a
multiple-source patch are unavailable, then a Required policy will throw
an error, otherwise simply return nil, moving on to processing the next
patch.

Signed-off-by: Ben Agricola <bagricola@squiz.co.uk>