Sync aws with vpc crossplane.yaml snippet with package
We are currently pinning the aws-with-vpc getting started package to
provider-aws v0.16.0 due to some API version changes that we have not
yet accomodated for. However, this is not reflected in the instructions
for building your own configuration package, so users who follow those
instructions will likely get a package that fails unless they already
have provider-aws v0.16.0 installed.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>