Create Lock if not exist in revision controller
Updates revision controllers to create the lock if it does not exist
when resolving dependencies. If a package revision is inactive it will
not create the lock as its only purpose would be to remove itself from
it. If a package opts out of dependency management it will not create
the lock as it does not attempt to consider dependencies.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>