Do not create resources if revision is inactive
Currently, if a revision is inactive and a resource in its package does
not exist in the cluster, it will try to create it and add itself as an
owner. This is typically not a problem, but if for some reason, such as
a CRD being updated in a subsequent revision and dropping the old
version, a resource must be deleted to allow a successful revision
upgrade, the old inactive revision and race with the new active one to
create the resource as it exists in their package. When the inactive
revision wins, the subsequent revision, which is active, will fail to
start because it cannot update the resource and add its controller
reference.

This changes inactive revisions to not create resources if they do not
exist, but still reconcile successfully. This means that if a resource,
such as a CRD gets deleted for an old inactive revision and it is not
present in the new active one, the old inactive revision will not create
that CRD and it will still report healthy. However, if the inactive
revision is rolled back to, it would recreate that resource before
starting its controller. In reality, it should never be a problem that
an inactive revision cannot create a missing resource type because a
user would have had to manually delete it, but this does unblock the
case where a user wants to delete the resource because it is preventing
upgrade.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>