Move extra env vars documentation from template to README
Removes the examples in the helm values template file and updates the
documentation to specify how extra environment variables can be
provided. This makes it such that the examples will be shown on the
documentation website and makes it such that we don't have some fields
documented in the template file but not others.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>