Use original package source string when adding and removing self from lock
Updates the package revision controller to use the original package
source as the source in the package lock so that configuration packages
are able to accurately match their dependencies in the lock. This is
necessary because go-containerregistry will overwrite docker.io to
index.docker.io to accomodate legacy docker format, which is desirable
when pushing or pulling images, but not when persisting the package
source and attempting to match it with another package's declared
dependencies.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>