Use go-containrregistry default registry if none provided
Makes the default registry match the upstream go-containerregistry
default if none is supplied. This same behavior could be accomplished by
not passing the WithDefaultRegistry option to places where we parse
image references, but it is a bit cleaner to always pass the option and
just populate it with the upstream default if not overridden with a flag
or environment variable. It also makes it more discoverable at the core
crossplane command level.

This change makes the default behavior match that from before the
default registry was configurable.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>