Annotate potentially misleading design docs
We have a few design documents that are still mostly accurate and interesting
when it comes to describing problems we've wanted to solve, but that either
mention defunct concepts (e.g. resource classes) or go into detail about the
initial rather than the current technical implementation of a problem (e.g. the
code examples in the cross resource references document).

Signed-off-by: Nic Cope <negz@rk0n.org>