Allow user control over pod labels using controllerconfig
To be able to select pods using PodSecurityPolicy in EKS, we need to
have predictable labels on the pods. Currently only an unpredictable
hash label is provided on the pods, and the user can only add
annotations.

This change adds a label map for the pod template, similar to the
existing annotation map.

To ensure future extensibility labels with crossplane.io prefixes are
forbidden in the documentation, and the code is structured to ensure
that the template will match the pod selector, and that any user
provided labels which overlaps with the selector will be overriden.

Fixes #2398

Signed-off-by: Carl Henrik Lunde <chlunde@ifi.uio.no>