Put CompositionRevision support behind a feature flag
This commit introduces a simple feature flag library and uses it to gate
CompositionRevision support. Going forward I'd like to keep all newly introduced
features of any significant size and complexity behind a feature flag like this.

The idea here is that the feature flags are a simple map that can be plumbed
down from cmd/ to the various places in the codebase we'd need to use them. This
approach mostly just avoids having to plumb down a lot of boolean flags as
distinct bool types, which would bloat our controller Setup() signatures in
particular.

Signed-off-by: Nic Cope <negz@rk0n.org>