Use GenerateName, not revision number for CompositionRevision names
This reverts to the original design here. It was suggested that we use the
revision number as a suffix instead during review. I like that approach, but
have since realised we're using GenerateName for package revisions. Given it
does not make a big difference (the revision number appears prominently in the
spec and as a printer column) I prefer to keep symmetry with pacakge revisions.

Signed-off-by: Nic Cope <negz@rk0n.org>