Use the hash, not revision, to determine the latest Composition
This is necessary for the Automatic update policy to actually track the 'latest'
(i.e. most current) revision. We cannot rely on the revision number, because
it's possible that the latest Composition state may actually be a reversion to
an older state, and thus an older (lower numbered) CompositionRevision. Matching
the hash ensures we always select the revision that matches the current
Composition.

Signed-off-by: Nic Cope <negz@rk0n.org>