Remove Composition status
This status was added by accident, and has never been used. It doesn't really
hurt, but it's a little misleading given that a Composition never has status
conditions in practice.

Removing this is technically a breaking API change, but none of our controllers
have ever set it so I can't imagine anyone is actually depending on it.

Signed-off-by: Nic Cope <negz@rk0n.org>