Use embedded type metadata, not ObjectTypers, in claim Reconciler
I believe this pattern of using ObjectTypers was copied from a past iteration
of the Claim reconciler that dealt in strongly typed Claims. At some point those
types did not have their APIVersion and Kind deserialized when read from the API
server so we needed to look them up in the scheme.

This iteration of the claim reconciler always deals with types that are built on
the *unstructured.Unstructured type, which always has its APIVersion and Kind
set when it is read from the API server. Therefore we can just use GetObjectKind
rather than needing to pass around an ObjectTyper.

Signed-off-by: Nic Cope <negz@rk0n.org>
(cherry picked from commit 26050ab988581c0f57885051b9a423c480d07673)