Use a dry-run create to name XRs during Configure
This is similar to the approach we take in the XR reconciler when creating
composed resources. The dry-run create causes the XR's metadata.name to be
populated based on its metadata.generateName, without the XR actually being
created. This gives us a generated XR name that is very likely to be available
that we could persist as our claim's resourceRef _before_ we actually create the
XR, thus allowing us to reduce the likelihood that we would leak XRs.

Signed-off-by: Nic Cope <negz@rk0n.org>
(cherry picked from commit 5f9571c4665881707371e82e379e03cee6e9bdd7)