Persist claim's resourceRef before creating its XR
This commit leverages our newly implemented ability to name XRs before they're
created by persisting the XR's name to the claim's resourceRef before we create
(i.e. apply) the XR. 'Binding' a claim to an XR is now broken up into three
stages.

1. We set the XR's claimRef during Configure.
2. We set (and persist) the XRC's resourceRef during Bind
3. We persist the XR's claimRef during Apply.

This approach has a handful of benefits:

1. It prevents 'leaking' XRs. Previously if anything went wrong between when we
   first created an XR and when we persisted its resourceRef we'd return from
   the Reconcile and requeue a new one. The new reconcile would have no way of
   knowing that we'd already created an XR for this claim so it would create a
   new one.
2. It ensures dynamically provisioned XRs never exist without a claimRef - it's
   set when the XR is created. This can help avoid issues like #2464. Note that
   XRs should still be designed to support static binding - i.e. they should not
   assume that the claimRef will always be set at XR creation time.
3. It eliminates one superfluous API server write, because we only update the XR
   once, rather than applying it then updating it to bind it.

Note that this commit changes the behaviour of the (AFAIK _exceedingly_ obscure)
edge case in which someone attempts to claim an XR tha does not exist - i.e. by
explicitly setting their claim's resourceRef. Previously we'd assume they did in
fact want to claim an XR that hadn't yet been created, and wait around for it to
exist. Now we assume that the resourceRef was populated by a previous iteration
of the reconciler and proceed to dynamically provision an XR of that name. One
interesting side effect here is that this means claim authors can now influence
the (Kubernetes) name of the XR they would like to provision, though we'll still
bail out if they try to 'claim over' an XR that is claimed by some other entity.

Signed-off-by: Nic Cope <negz@rk0n.org>
(cherry picked from commit 066c20c787842a78b660dc561ada6523f839c31d)