'Future proof' the --enable-composition-revisions flag
This removes the 'alpha' part of the flag so that we can continue to use the
same flag once the feature reaches beta (when we'll have it on be default).
We instead use a kong flag group to highlight that the feature is currently in
alpha - in future we can move it into a beta group without a breaking change.

I believe we can also make the flag on by default by adding default:"true" and
negatable:"" in future.

Signed-off-by: Nic Cope <negz@rk0n.org>