Indicate which CompositionRevision is 'current'
The current CompositionRevision is the one with a spec that matches the
Composition's spec at any point in time. There should only ever be one
current CompositionRevision, and it won't necessarily be the one with the
highest revision number, e.g. if the Composition has reverted to a state that
matches a prior revision.

Signed-off-by: Nic Cope <negz@rk0n.org>