Add a guide to CompositionRevisions
I imagine this will be rolled into the broader Composition documentation at some
point, and probably framed more around the actual use case it solves, i.e.
'updating XRs' or 'evolving Compostions'. However for the time being while the
feature is in alpha I'd like to keep this documentation separate and frame it as
a guide to the specific feature.

Signed-off-by: Nic Cope <negz@rk0n.org>