Capture common style feedback left during code review
We've historically avoided having a 'style guide' for the Crossplane project,
instead deferring to various excellent upstream Go documents like 'Effective
Go', and 'Go Code Review Comments'.

Unfortunately these documents are quite a lot to get through when put together,
so it's unrealistic to expect most contributors to digest them all. In this
commit I've attempted to update CONTRIBUTING.md to provide a non-exhaustive list
of the style comments I find myself most frequently leaving on Crossplane pull
requests. My hope is that the community might find this to be a little more
discoverable, or that it will at least give us an easy place to point folks when
we're reviewing code.

Signed-off-by: Nic Cope <negz@rk0n.org>