Strip version from package name when installed with initializer
Updates the initializer package installer to not include version in name
used for rendered Provider or Configuration. This fixes the scenario in
which a user upgrades Crossplane and specifies new versions of existing
packages, which would previously cause a new package installation,
rather than just an upgrade.

Note that this change would cause a user who previously used the
initializer to install packages with an existing Crossplane version to
get duplicates on upgrade even if they did not change the package
versions since the new package installs would have no version
information. However, this does improve the experience of future
upgrades after the initial one. Users who start with a Crossplane
version that includes this fix should not experience any duplicate
installs on future upgrades.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>