Switch to lease based leader election
The default in controller-runtime is to use the 'configmapsleases' strategy,
which takes a lock using both a ConfigMap and a Lease. This is done for backward
compatibility. For reasons that are not immediately clear to me (presumably some
kind of saturation) we find that at --max-reconcile-rate=100 we start to see
leadership losses because we can't renew our lease before the 10 second renewal
deadline is up. This should be a safe change since we've been using
configmapleases for many releases, so I don't expect dropping configmaps to
affect anyone. (We also don't create more than one replica by default.)

Signed-off-by: Nic Cope <negz@rk0n.org>