Expose provider Prometheus metrics by default
This makes it easier to use pod/service based Prometheus service discovery
without needing to provide a ControllerConfig. In particular it's not possible
to use a prometheus-operator PodMonitor or ServiceMonitor to discover targets
without a named port to target. I feel this is a common enough use case that it
makes sense to be on by default or all providers rather than requiring a
ControllerConfig.

Signed-off-by: Nic Cope <negz@rk0n.org>