Label provider deployments with provider name
In practice the name isn't that interesting/useful - the goal is mostly to add a
label (pkg.crossplane.io/provider) that can be used to find all deployments/pods
that are Crossplane providers. Technically we could look for anything with the
pkg.crossplane.io/revision label, but provider is more explicit/intuitive.

Signed-off-by: Nic Cope <negz@rk0n.org>