Check existing installed packge from matching source in installer
Updates the initializer package installer to check if a package already
exists with the provided source, and update the version without changing
the name if it does. This makes it possible to update package versions
on Crossplane install even if the Provider or Configuration object name
was not previously set by the installer or was set by a previous version
of the installer that used a different naming strategy.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>
(cherry picked from commit 50141b39e49e621829389c764e0725b76281be2a)