Switch to using Packages and Configurations directly for matching pkg
Updates the initializer installer to use list of configurations and
packages directly to get matching package sources. This is a better
indication of the presence of a package that should have its version
updated than the lock because its possible that a package could be in
process of having its revision add itself to the lock, or may be in some
error state that it can't add itself to the lock. We still want to
update versions in each of those cases.

Also note that we explicitly ignore packages with invalid references in
the package source. This indicates that a package was pre-loaded in the
cache and we can't make assumptions about updating it to a newer
version.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>
(cherry picked from commit f9ed58c76639db389e308f4e993a012ebf66c0ce)