Copy all binaries into image temp dir and select in Dockerfile
Updates the image build makefile to copy all binaries into the image
temp directory so that they can be selectively added for the appropriate
platform during a multi-platform build and push (i.e. make publish). The
built-in TARGETOS and TARGETARCH variables are used to select the binary
for the apprpriate platform in the Dockerfile.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>