Update Azure credentials step to not require AD access
The permissions we grant in this step are highly privileged and are
currently only required if we are provisioning AKS clusters, which we no
longer do in the getting started guide. Many users will not be able to
set the permissions, and may be deterred from continuing if they
encounter an error. This removes them as they are not required for the
operations performed in subsequent steps.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>
(cherry picked from commit 8eaebd17b1bbf84b6c0932b61f56eaf46a8a4b4e)