Only set package revision in scope of image backend Init
Modifies the ImageBackend by removing the member package revision and
instead using a local nestedBackend to hold the package revision locally
to the Init method. This fixes a bug where simultaneous reconciles of
providers or configurations could overwrite the package revision
mid-init, causing a given revision to install and control the wrong
types.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>