Patch multiple array fields when passed a wildcarded path
This change allows patches of the FromCompositeFieldPath and
ToCompositeFieldPath types to patch from one field in the "from"
resource to multiple object fields nested inside an array in the "to" resource.
This allows an indefinite number of arrays to be patched without
needing to specify a patch per each individual element.

Signed-off-by: EdgeJ <5093048+EdgeJ@users.noreply.github.com>