Make release template (arguably) less onerous
When running a release I prefer to be able to simply check off boxes via
the GitHub UI (i.e. rendered, not in markdown). I find that having to
keep the issue open in markdown mode while I insert links to each step
to be quite onerous. In practice GitHub automatically cross references
most things (e.g. PRs and commits) as long as they're opened with a
reference back to the release issue. Anything else (e.g. CI runs, Tweets
etc) would (in my opinion) be quicker to drop in as a comment on the
release issue.

Signed-off-by: Nic Cope <negz@rk0n.org>