Offer claim in composition concept docs
Updates the XRD in the composition concept docs to offer a claim as we
later show what a claim for an XR would look like. We aren't explicitly
saying that the claim could be created with the manifest of the XRD as
it currently stands, but it can be confusing for users that we show an
example of a claim when the XRD in the same doc does not actually offer
a claim type.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>