Update image backend for parser to not depend on cache
Refactors the image parser backend to not depend on cache and instead
only deal with pulling and extracting YAML streams from images. The
cache will now be managed separately from the image backend, and the
image backend will only be called in the event we don't have the data in
the cache.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>