feat(compositions): Support patching from environment
This adds a new cluster scope CRD `EnvironmentConfig` that is similar to a
standard `ConfigMap` but allows storing arbitrary values instead of just
strings.

EnvironmentConfigs can be used in compositions using the new `FromEnvironmentFieldPath`
patch type.

The main use case of this PR is to use to build a generic Crossplane
package that can run in multiple environment without having to hardcode
environment specific values in the compositions.

It avoids the discussed security and RBAC issues by not requiring access
to secrets and config maps and relying on Crossplane native resources
instead.

This is designed to match the design spec in
[`./design/one-pager-composition-environment.md`](./design/one-pager-composition-environment.md].

Signed-off-by: Maximilian Blatt <maximilian.blatt-extern@deutschebahn.com>
(external expert on behalf of DB Netz AG)