Move bassam and jbw976 to emeritus maintainers
In practice bassam and jbw976 haven't been actively merging code or
grooming issues in some time, so we'd like to reflect that by moving
them to emeritus maintainers. Both are still members of the steering
committee, and thus continue to have admin access to this repository
(and all others in the github.com/crossplane org).

Signed-off-by: Nic Cope <nicc@rk0n.org>