Prevent flux from accidentally deleting serviceaccount when using ControllerConfig
Flux will garbage collect the ServiceAccount if we copy the flux
annotations from the ControllerConfig.

Background:

We sometimes see bursts of Unauthorized for calls from crossplane
providers to the kubernetes apiserver. After some debugging we've
found that the UID in the service account does not match the token on
the file system. We noticed that the service account has a "fluxcd.io"
annotation, even though it is not managed by flux. This is because
crossplane copies all annotations from the `ControllerConfig` to the
`ServiceAccount` object. This is probably required for ArgoCD, and it
is used for IRSA in AWS.

But for flux, this does not work. We believe flux will consider this
as a situation where a former `Kustomization` managed this object, and
delete it as it is no longer managed. crossplane will quickly recreate
it, but the token mounted in the container will for some time have the
old UID and the API server will reject the requests. I'm not sure if
this is a race condition in the kubelet when a token is rotated while
a pod is created, or if it is simply not a supported way to rotate the
tokens by deleting the service account while it is in use, but
Kubernetes does recover from this situation after a while.

Fixes #3030

Signed-off-by: Carl Henrik Lunde <chlunde@ifi.uio.no>