Add a project charter
The Crossplane steering committee hopes that establishing a charter will help
our community better understand our intentions for the project - in particular
what is and isn’t in scope. As a byproduct of this, we hope it will help
vendors better understand where Crossplane ends and they start.

The document is informed by the CNCF’s guidance [1] around project charters,
which they recommend consist of a brief mission statement, supplemented by
scope and principles. It’s also heavily inspired by the “What is Kubernetes?”
page [2] (which appears to be the closest thing the Kubernetes project as a
whole has to a charter). Like said document this one attempts to describe what
Crossplane is for and why it exists without using or introducing specific
Crossplane terminology (e.g. “provider”, XR, etc).

1: https://contribute.cncf.io/maintainers/governance/charter/
2: https://kubernetes.io/docs/concepts/overview/what-is-kubernetes/

Signed-off-by: Nic Cope <nicc@rk0n.org>