Switch back to alpine
Unfortunately getting crun to work with distroless is a massive pain.
The static builds available on their GitHub releases don't seem to work
(I get an OOM and an error about an invalid syscall - I forget which).
We could build a dynamically linked crun, but it's unclear how to add
the libs that it would need to run to distroless without a package
manager (and without using bazel).

Alpine is as small or smaller than distroless, and has a relatively
modern crun right there in its package manager. As far as I can tell the
only downside of returning to alpine is the larger attack surface that
comes with having busybox and apk available.

Signed-off-by: Nic Cope <nicc@rk0n.org>