Don't link to latest patch releases in README.md
I'm sure it's of non-zero use, but I'm (very) skeptical
that it's worth the extra step in the release process.

GitHub releases may not have the ideal UX for a project
that releases patches on multiple release branches but
it's quite a common pattern nonetheless and I don't see
too many other projects going out of their way to clarify
which one is latest. Speaking personally (as a user of
other projects) I'd find this to be a minor convenience
at best.

Signed-off-by: Nic Cope <nicc@rk0n.org>