:bug: add new composition for new AWS VPC
We can't use `claim-aws.yaml` to create RDS with new VPC because we need `vpc: new`.

I add new file and fix doc.
I was able to change the procedure to edit the file.
However, I wanted to just apply the file like the other example, so I created a new file.

I applied k8s cluster on my PC and work it.

Signed-off-by: ota42y <ota42y@gmail.com>