Ensure the stale workflow uses fewer operations
We only get 1,000 operations per hour for every workflow that uses
GITHUB_TOKEN. We're now seeing the stale workflow use relatively few
operations with few to no stale issues open.

In addition to lowering the workflow's operation limit this change has
the workflow run once a day. It adds a new `/fresh` comment command that
can be used to instantly mark an issue or PR as not stale, rather than
waiting up to a day for the stale workflow to run.

Signed-off-by: Nic Cope <nicc@rk0n.org>