Drop NodeFns and FindIndex from DAG interface
Passing NodeFns proved to not be a very ergonomic API to work with and
most places where FindIndex was used could be replaced by either working
with the Nodes directly or calling NodeExists. The index of a Node
should not be the responsibility of the DAG, which provides an interface
into a graph.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>