Use package revision name to identify self in Lock
Updates the package revision dependency manager to use the package
revision name to identify self in the Lock. This prevents the case of
one package revision attempting to remove itself from the Lock while
another is adding with the same source. Source is still used as an
identifier in the DAG as it dictates whether we determine a dependency
is present or not.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>
(cherry picked from commit cdcf68f521fcb41bcb0851b199e6d445b04be81f)