Use package revision name to identify self in Lock
Updates the package revision dependency manager to use the package
revision name to identify self in the Lock. This prevents the case of
one package revision attempting to remove itself from the Lock while
another is adding with the same source. Source is still used as an
identifier in the DAG as it dictates whether we determine a dependency
is present or not.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>