Add note recommending usage of SPDX License Identifiers
Adds a note to the xpkg specification recommending that an SPDX License
Identifier be used for the license annotation. This is not strictly
required because it is not expected that tooling will reject a package
that uses a non-valid SPDX License Identifier.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>
(cherry picked from commit b397b62914c3aa162b0abf24a923490b3ad19f72)