Switch from ResourceList to FunctionIO
The design of the ResourceList type was heavily inspired by KRM
functions, but the implementation proposed by the Composition Functions
design was already incompatible (see "Using KRM Spec Compatible
Functions" under "Alternatives Considered"). It also recently occurred
to me that Crossplane may need information that isn't easily modelled as
part of a simple array of Kubernetes resources - for example it may need
to know how to derive an XR's connection details from its composed
resources'.

To this end it seemed better to break away completely from ResourceList
and design a similar type that was purpose built for Composition
Functions. Some improvements in this type include:

* Better schema for `composite` - we can ensure exactly one is required,
  and easily find it without iterating through the list of resources.
* We can use real (required) fields in the resources array, not special
  annotations, for things like the name of the array entry.
* We can use fields in the resources array to return connection details
  (or instructions on how to derive connection details)

Signed-off-by: Nic Cope <nicc@rk0n.org>