Break FunctionIO into 'observed' and 'desired' sections
https://github.com/crossplane/crossplane/pull/2886#discussion_r991710579

The above comment thread contains the discussion that lead to this
change. The primary motiviation for this change is that sometimes we'll
want to provide functions with context (observed state) that they can't
mutate, and we want to clearly distinguish that from state that they can
mutate.

A simple example is the status of composed resources - functions may use
this status as an input, but can't mutate it. Connection details are a
more complex example, in that the format used to specify desired
connection details differs from the format used to surface existing,
observed connection details.

This approach has some additional benefits:

* All functions are able to know the observed state of the world, which
  may differ from the accumulated desired state.
* Desired state may be returned in a more succinct 'overlay' style. This
  may be more or less onerous than mutating observed state depending on
  the scenario.
* Passing observed state (without accidentally mutating it) is likely to
  be less error prone.

Signed-off-by: Nic Cope <nicc@rk0n.org>