Have the revive linter act like golint
At some point in recent history https://github.com/golang/lint was
archived, and is now disabled by default in golangci-lint. We relied on
this linter to - among presumably other things - ensure that all
exported symbols had a docstring. This rule is not currently being
enforced.

https://revive.run is intended to be a drop-in replacement for golint.
It's already enabled in our golangci-lint config, but it seems that
supplying _any_ configuration for it overrides its default behaviour,
which is to match golint's configuration.

I do not that when in golint mode it's enforcing a package-doc rule that
golint didn't seem to be enforcing.

Signed-off-by: Nic Cope <nicc@rk0n.org>