Automatically replicate identical API types to older versions
Often we'll promote a type from (e.g.) v1alpha1 -> v1beta1, using the
'none' conversion strategy that relies on the two types remaining the
same. This has historically caused issues when someone updates one type
(typically the newer one) and forgets to update the other.

In order to avoid this issue I've added a simple script that 'generates'
identical older types from the latest type (which should be the storage
version).

Signed-off-by: Nic Cope <nicc@rk0n.org>